Introduce a shared scm context
In some situations, when many local cookbooks are used, tests can spend a lot of
time doing several times the same git commands for each one of them. It can be
particularly inefficient when they share the same git repo.

With this change, we make sure commands are launched only once and results are
reused if necessary.